chap7/requirements.txt
google-generativeai
python-dotenv
streamlit
yfinance
pytz
matplotlib
pandas
beautifulsoup4
requests
tabulate
